Linear Search in Arrays
Linear search in C programming: The following code implements linear
search (Searching algorithm) which is used to find whether a given
number is present in an array and if it is present then at what location
it occurs. It is also known as sequential search. It is straightforward
and works as follows: We keep on comparing each element with the element
to search until it is found or the list ends.

Signed-off-by: Panchajanya1999 <rsk52959@gmail.com>